+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(kwb.raindrop)
+
+test_check("kwb.raindrop")